+++ NEWS.md
@@ -1,10 +1,11 @@
-# lava 1.8.3 (development version)
-  - `sim.default`: the return object can be an `estimate` object
+# lava 1.8.3
+  - `sim.default`: the simulation function `f`'s return object can now be an
+    `estimate` object
   - `summary.sim`: automatically derives estimate, se, confint parameters if the
      sim routine returns an `estimate` object
   - `merge.estimate`: cast warning if `back.transform` was used
   - `merge.estimate`: works with objects with and without influence function
-  - `summary.sim` df argument for calculating CIs based on t-dist. approximation
+  - `summary.sim` `df` argument for calculating CIs based on t-dist. approximation
 
 # lava 1.8.2
   - Improved closed testing procedure `closed_testing` (depr. `closed.testing`)